utils/text_extraction.py
# You may need libraries such as PyPDF2 for PDFs, python-docx for docx, etc.
import PyPDF2
import docx

from PyPDF2 import PdfReader

def extract_text_from_cv(file_stream):
    # Assuming the uploaded file is a PDF
    try:
        reader = PdfReader(file_stream)
        text = ""
        for page in reader.pages:
            text += page.extract_text()
        return text
    except Exception as e:
        raise e  # Re-raise the exception to be caught by the calling function

def extract_job_description_from_link(link):
    # Placeholder for extracting job description from a link
    # You would use requests, BeautifulSoup, etc. to scrape content from the link
    # Be sure to handle exceptions and errors.
    pass